+++ entity/src/epilot/models/operations/createsavedview.py
@@ -2,17 +2,15 @@
 
 from __future__ import annotations
 import dataclasses
-import requests as requests_http
-from ..shared import savedviewitem as shared_savedviewitem
+from ...models.components import httpmetadata as components_httpmetadata
+from ...models.components import savedviewitem as components_savedviewitem
 from typing import Optional
 
 
 @dataclasses.dataclass
 class CreateSavedViewResponse:
+    http_meta: components_httpmetadata.HTTPMetadata = dataclasses.field()
+    saved_view_item: Optional[components_savedviewitem.SavedViewItem] = dataclasses.field(default=None)
+    r"""Success"""
     
-    content_type: str = dataclasses.field()  
-    status_code: int = dataclasses.field()  
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)  
-    saved_view_item: Optional[shared_savedviewitem.SavedViewItem] = dataclasses.field(default=None)
-    r"""Success"""  
-    
+